+++ IBBS/Pages/BurgerPage.cshtml
@@ -3,71 +3,54 @@
 @{
     ViewData["Title"] = "Burgers";
 }
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"] </title>
-    <link rel="stylesheet" href="~/css/burgerpage.css" />
-    <link rel="stylesheet" href="~/lib/font-awesome/css/all.css" />
-</head>
-<body>
-    <ol>
-        @foreach (var item in Model.Ingredients)
-        {
-            <li>@item.Name</li>
-        }
-    </ol>
-    <div id="heading" class="container-sm">
-        <h1>Burgers</h1>
-    </div>
-    <div id="burgercarousel" class="carousel slide" data-ride="carousel">
-        <ol class="carousel-indicators">
-            <li data-target="#burgercarousel" data-slide-to="0" class="active"></li>
-            <li data-target="#burgercarousel" data-slide-to="1"></li>
-            <li data-target="#burgercarousel" data-slide-to="2"></li>
-        </ol>
-        <div class="carousel-inner">
-            <div class="carousel-item active">
-                <img src="~/images/IBBSburger.png" class="d-block w-100" alt="...">
-                <div id="picturetext" class="carousel-caption d-none d-md-block">
-                    <h5>First slide label</h5>
-                </div>
-            </div>
-            <div class="carousel-item">
-                <img src="~/images/IBBSburger.png" class="d-block w-100" alt="...">
-                <div id="picturetext" class="carousel-caption d-none d-md-block">
-                    <h5>Second slide label</h5>
-                </div>
-            </div>
-            <div class="carousel-item">
-                <img src="~/images/IBBSburger.png" class="d-block w-100" alt="...">
-                <div id="picturetext" class="carousel-caption d-none d-md-block">
-                    <h5>Third slide label</h5>
-                </div>
-            </div>
-        </div>
-        <a class="carousel-control-prev" href="#burgercarousel" role="button" data-slide="prev">
-            <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-            <span class="sr-only">Previous</span>
-        </a>
-        <a class="carousel-control-next" href="#burgercarousel" role="button" data-slide="next">
-            <span class="carousel-control-next-icon" aria-hidden="true"></span>
-            <span class="sr-only">Next</span>
-        </a>
-    </div>
-    <div id="likes" class="container">
-        <div class="row justify-content-md-center">
-            <div class="col-lg-2">
-                <i class="far fa-thumbs-up"></i>
-            </div>
-            <div class="col-lg-2">
-                <i class="far fa-thumbs-down"></i>
-            </div>
-        </div>
-    </div>
 
-    
-</body>
-</html>
+<div id="heading" class="container-sm">
+    <h1>Burgers</h1>
+</div>
+<div id="burgercarousel" class="carousel slide" data-interval="false" data-ride="carousel">
+    <ol class="carousel-indicators">
+        <li data-target="#burgercarousel" data-slide-to="0" class="active"></li>
+        <li data-target="#burgercarousel" data-slide-to="1"></li>
+        <li data-target="#burgercarousel" data-slide-to="2"></li>
+    </ol>
+    <div class="carousel-inner">
+        <div class="carousel-item active">
+            <img src="~/images/IBBSburger.png" class="d-block w-100" alt="...">
+            <div id="picturetext" class="carousel-caption d-none d-md-block">
+                <h5>First slide label</h5>
+            </div>
+        </div>
+        <div class="carousel-item">
+            <img src="~/images/IBBSburger.png" class="d-block w-100" alt="...">
+            <div id="picturetext" class="carousel-caption d-none d-md-block">
+                <h5>Second slide label</h5>
+            </div>
+        </div>
+        <div class="carousel-item">
+            <img src="~/images/IBBSburger.png" class="d-block w-100" alt="...">
+            <div id="picturetext" class="carousel-caption d-none d-md-block">
+                <h5>Third slide label</h5>
+            </div>
+        </div>
+    </div>
+    <a class="carousel-control-prev" href="#burgercarousel" role="button" data-slide="prev">
+        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+        <span class="sr-only">Previous</span>
+    </a>
+    <a class="carousel-control-next" href="#burgercarousel" role="button" data-slide="next">
+        <span class="carousel-control-next-icon" aria-hidden="true"></span>
+        <span class="sr-only">Next</span>
+    </a>
+</div>
+<div id="likes" class="container">
+    <div class="row justify-content-md-center">
+        <div class="col-lg-2">
+            <span id="Likes">420</span>
+            <img class="thumbspic thumbup" id="thumbsup" src="~/images/Tummeup.png" />
+        </div>
+        <div class="col-lg-2">
+            <img class="thumbspic thumbdown" id="thumbsdown" src="~/images/Tummener.png" />
+            <span id="Dislike">69</span>
+        </div>
+    </div>
+</div>
